Added basic FAILED address correction method
This update improves the geocoding of the test CSV from 67.5 % to 75 %
of all addresses. It does so by taking address that failed to geocode
and trying to fix them by removing special characters, extra spaces, and
by adding "ST","ND","RD","TH" to numbered roads that are missing those
suffixes.